// cmd/web-chat/web/src/sem/pb/proto/sem/base/agent_pb.ts
// @generated by protoc-gen-es v2.11.0 with parameter "target=ts,import_extension=none"
// @generated from file proto/sem/base/agent.proto (package sem.base.agent, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import { file_google_protobuf_struct } from "@bufbuild/protobuf/wkt";
import type { JsonObject, Message } from "@bufbuild/protobuf";

/**
 * Describes the file proto/sem/base/agent.proto.
 */
export const file_proto_sem_base_agent: GenFile = /*@__PURE__*/
  fileDesc("Chpwcm90by9zZW0vYmFzZS9hZ2VudC5wcm90bxIOc2VtLmJhc2UuYWdlbnQiTwoLQWdlbnRNb2RlVjESCgoCaWQYASABKAkSDQoFdGl0bGUYAiABKAkSJQoEZGF0YRgDIAEoCzIXLmdvb2dsZS5wcm90b2J1Zi5TdHJ1Y3RCQlpAZ2l0aHViLmNvbS9nby1nby1nb2xlbXMvcGlub2NjaGlvL3BrZy9zZW0vcGIvcHJvdG8vc2VtL2Jhc2U7YmFzZWIGcHJvdG8z", [file_google_protobuf_struct]);

/**
 * @generated from message sem.base.agent.AgentModeV1
 */
export type AgentModeV1 = Message<"sem.base.agent.AgentModeV1"> & {
  /**
   * @generated from field: string id = 1;
   */
  id: string;

  /**
   * @generated from field: string title = 2;
   */
  title: string;

  /**
   * @generated from field: google.protobuf.Struct data = 3;
   */
  data?: JsonObject;
};

/**
 * Describes the message sem.base.agent.AgentModeV1.
 * Use `create(AgentModeV1Schema)` to create a new message.
 */
export const AgentModeV1Schema: GenMessage<AgentModeV1> = /*@__PURE__*/
  messageDesc(file_proto_sem_base_agent, 0);