// cmd/web-chat/web/src/sem/pb/proto/sem/base/debugger_pb.ts
// @generated by protoc-gen-es v2.11.0 with parameter "target=ts,import_extension=none"
// @generated from file proto/sem/base/debugger.proto (package sem.base.debugger, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import { file_google_protobuf_struct } from "@bufbuild/protobuf/wkt";
import type { JsonObject, Message } from "@bufbuild/protobuf";

/**
 * Describes the file proto/sem/base/debugger.proto.
 */
export const file_proto_sem_base_debugger: GenFile = /*@__PURE__*/
  fileDesc("Ch1wcm90by9zZW0vYmFzZS9kZWJ1Z2dlci5wcm90bxIRc2VtLmJhc2UuZGVidWdnZXIijAEKD0RlYnVnZ2VyUGF1c2VWMRIKCgJpZBgBIAEoCRIQCghwYXVzZV9pZBgCIAEoCRINCgVwaGFzZRgDIAEoCRIPCgdzdW1tYXJ5GAQgASgJEhMKC2RlYWRsaW5lX21zGAUgASgDEiYKBWV4dHJhGAYgASgLMhcuZ29vZ2xlLnByb3RvYnVmLlN0cnVjdEJCWkBnaXRodWIuY29tL2dvLWdvLWdvbGVtcy9waW5vY2NoaW8vcGtnL3NlbS9wYi9wcm90by9zZW0vYmFzZTtiYXNlYgZwcm90bzM", [file_google_protobuf_struct]);

/**
 * @generated from message sem.base.debugger.DebuggerPauseV1
 */
export type DebuggerPauseV1 = Message<"sem.base.debugger.DebuggerPauseV1"> & {
  /**
   * @generated from field: string id = 1;
   */
  id: string;

  /**
   * @generated from field: string pause_id = 2;
   */
  pauseId: string;

  /**
   * @generated from field: string phase = 3;
   */
  phase: string;

  /**
   * @generated from field: string summary = 4;
   */
  summary: string;

  /**
   * @generated from field: int64 deadline_ms = 5;
   */
  deadlineMs: bigint;

  /**
   * @generated from field: google.protobuf.Struct extra = 6;
   */
  extra?: JsonObject;
};

/**
 * Describes the message sem.base.debugger.DebuggerPauseV1.
 * Use `create(DebuggerPauseV1Schema)` to create a new message.
 */
export const DebuggerPauseV1Schema: GenMessage<DebuggerPauseV1> = /*@__PURE__*/
  messageDesc(file_proto_sem_base_debugger, 0);